updated sitemap

+++ packages/robust/core/src/Console/Commands/ResetWidget.php
@@ -40,29 +40,37 @@ class ResetWidget extends Command
      */
     public function handle(Filesystem $filesystem, Widget $widget)
     {
-        $packages = CoreHelper::names();
-        DB::table('widgets')->truncate();
+        $this->info("\n=============================================");
+        $this->info("========== Resetting Widgets Table ============");
+        $this->info("===============================================");
+        $execute = $this->confirm("Would you like to reset widgets table seeder? [y|N]", false);
         
-        foreach ($packages as $key => $package) {
-            $this->info("Entering {$package}");
-
-            $package_widgets = config("{$key}.widgets");
+        $packages = CoreHelper::names();        
+        if ($execute) {
+            $packages = CoreHelper::names();
+            DB::table('widgets')->truncate();
             
-            if (is_array($package_widgets)) {
-                foreach ($package_widgets as $package_widget) {
-                    if ($filesystem->exists(base_path() . "/packages/robust/{$key}/resources/views/admin/dashboard-widgets/{$package_widget['file_name']}.blade.php")) {
-                        $widget_data = [
-                            'name' => $package_widget['display_name'],
-                            'path' => $package_widget['file_name'],
-                            'package_name' => $key,
-                            'permission' => $package_widget['permission']
-                        ];
-                        $widget->create($widget_data);
+            foreach ($packages as $key => $package) {
+                $this->info("Entering {$package}");
+
+                $package_widgets = config("{$key}.widgets");
+                
+                if (is_array($package_widgets)) {
+                    foreach ($package_widgets as $package_widget) {
+                        if ($filesystem->exists(base_path() . "/packages/robust/{$key}/resources/views/admin/dashboard-widgets/{$package_widget['file_name']}.blade.php")) {
+                            $widget_data = [
+                                'name' => $package_widget['display_name'],
+                                'path' => $package_widget['file_name'],
+                                'package_name' => $key,
+                                'permission' => $package_widget['permission']
+                            ];
+                            $widget->create($widget_data);
+                        }
+
                     }
-
                 }
-            }
 
+            }
         }
     }
 }
